+++ metaboxes/metabox.php
@@ -77,10 +77,10 @@ class STM_Metaboxes {
 
 					if ( isset( $_POST[ $field_name ] ) ) {
 
-						if ( 'editor' === $field['type'] || 'curriculum' === $field['type'] ) {
-							$field_modified = ( is_array( $_POST[ $field_name ] ) ) ? filter_var_array( $_POST[ $field_name ] ) : $_POST[ $field_name ];
+						if ( 'editor' === $field['type'] ) {
+							$field_modified = ( is_array( $_POST[ $field_name ] ) ) ? filter_var_array( $_POST[ $field_name ] ) : $_POST[ $field_name ]; // phpcs:ignore
 						} else {
-							$field_modified = ( is_array( $_POST[ $field_name ] ) ) ? filter_var_array( $_POST[ $field_name ], FILTER_SANITIZE_STRING ) : sanitize_text_field( $_POST[ $field_name ] );
+							$field_modified = ( is_array( $_POST[ $field_name ] ) ) ? filter_var_array( $_POST[ $field_name ], FILTER_SANITIZE_STRING  ) : sanitize_text_field( $_POST[ $field_name ] ); // phpcs:ignore
 						}
 
 						if ( method_exists( 'STM_Metaboxes', "wpcfto_field_sanitize_{$field['type']}" ) ) {
@@ -739,8 +739,8 @@ function wpcfto_metaboxes_display_group_field( $section, $section_name, $field, 
 		<div class="row">
 		<?php if ( isset( $field['group_title'] ) && ! empty( $field['group_title'] ) ) { ?>
 		<div class="wpcfto_group_title"><?php echo esc_html( $field['group_title'] ); ?></div>
-	<?php } ?>
-	<?php
+			<?php
+		}
 	endif;
 
 	wpcfto_metaboxes_display_single_field( $section, $section_name, $field, $field_name );
@@ -748,7 +748,7 @@ function wpcfto_metaboxes_display_group_field( $section, $section_name, $field, 
 	if ( 'ended' === $field['group'] ) :
 		?>
 		</div></div></div>
-	<?php
+		<?php
 	endif;
 }
 
@@ -757,8 +757,7 @@ function wpcfto_metaboxes_preopen_field( $section, $section_name, $field, $field
 	$field_options   = $section['fields'][ $field_name ];
 	$preopen_disable = '';
 	?>
-	<div class="preopen_field_wrapper wpcfto_generic_field"
-		 v-init="initOpen(<?php echo wp_kses( $vue_field, array() ); ?>)">
+	<div class="preopen_field_wrapper wpcfto_generic_field" v-init="initOpen(<?php echo wp_kses( $vue_field, array() ); ?>)">
 		<?php
 		if ( ! stm_wpcfto_is_pro() && isset( $field_options['disabled'] ) && true === $field_options['disabled'] ) {
 			$preopen_disable = 'preopen_disable';
@@ -769,8 +768,7 @@ function wpcfto_metaboxes_preopen_field( $section, $section_name, $field, $field
 
 			<label class="<?php echo esc_attr( $preopen_disable ); ?>">
 
-				<div class="wpcfto-admin-checkbox-wrapper"
-					 v-bind:class="{'active' : <?php echo esc_attr( $vue_field ); ?>['opened'], 'is_toggle' : <?php echo ( isset( $field['toggle'] ) ) ? esc_attr( $field['toggle'] ) : 'true'; ?>}">
+				<div class="wpcfto-admin-checkbox-wrapper" v-bind:class="{'active' : <?php echo esc_attr( $vue_field ); ?>['opened'], 'is_toggle' : <?php echo ( isset( $field['toggle'] ) ) ? esc_attr( $field['toggle'] ) : 'true'; ?>}">
 					<div class="wpcfto-checkbox-switcher"></div>
 				</div>
 
@@ -780,8 +778,7 @@ function wpcfto_metaboxes_preopen_field( $section, $section_name, $field, $field
 
 		</div>
 
-		<div class="preopen_field"
-			 v-if="<?php echo esc_attr( $vue_field ); ?>['opened']">
+		<div class="preopen_field" v-if="<?php echo esc_attr( $vue_field ); ?>['opened']">
 			<?php wpcfto_metaboxes_display_single_field( $section, $section_name, $field, $field_name ); ?>
 		</div>